social-otter/app/requirements.txt
requests
snscrape
firebase_admin
pandas
sentry-sdk
pydantic
python-dotenv
flake8
# ML Dependencies
transformers
numpy
scipy
torch
# Utils
rich
